Prevent merging conference calls hosted on peer device.
With IMS, the "multiparty" bit on an ImsCall is set to "true" when a call
is merged into a conference.  This not only occurs on the device hosting
the conference call, but also on the devices of the callers merged into
the conference.

This CL adds code to ImsCall to handle changes to the multiparty state of
the session.  It also includes code to track whether the ImsCall is the
Conference Host.  A call is the conference host if the conference was
established on the current device.

Bug: 19478784
Change-Id: I764d6381d25275980a87705b7b53e21cdbffb1ea